I added commenting to files to help make them more understandable.
I also made initDB.js work. The issue was that I had forgotten to change the model name from 'Project' to 'Story'. The
DB should be able to really start working now.

I also had to make a new json file to work with their initDB script (their script used an array, and we were/are using a
map version because that's easier. The default stories.json we can leave as an array now since we don't need to access it
and all we have to do is just read from it when we reset the server.